requirements.txt
alembic==1.15.2
Sqlalchemy==2.0.37
python-dotenv==1.0.1
argon2-cffi